requirements.txt
requests
python-dotenv
pillow
streamlit

